split configuration into submodule
